add protocol: modify MAC; fix bug: heartbeat timeout 0xff dead-loop

+++ nc600-2.0/src/client.c
@@ -13,7 +13,7 @@
 
 
 u8 heartbeat_s =HEARTBEAT_TIMEOUT;
-u8 heartbeat_timeout = 0;
+u16 heartbeat_timeout = 0;/*FIX BUG001:*/
